Refactor the package seedu.address.model.person to seedu.address.model.order.
