+++ src/main/java/seedu/address/logic/commands/AddCommand.java
@@ -9,7 +9,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_TAG;
 
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.Order;
 
 /**
  * Adds a person to the address book.
