+++ src/main/java/seedu/address/model/OrderBook.java
@@ -5,8 +5,8 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Order;
-import seedu.address.model.person.UniqueOrderList;
+import seedu.address.model.order.Order;
+import seedu.address.model.order.UniqueOrderList;
 
 /**
  * Wraps all data at the order-book level
