+++ src/main/java/seedu/address/model/ReadOnlyOrderBook.java
@@ -1,7 +1,7 @@
 package seedu.address.model;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.Order;
 
 /**
  * Unmodifiable view of an order book
