+++ src/main/java/seedu/address/model/order/NameContainsKeywordsPredicate.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.order;
 
 import java.util.List;
 import java.util.function.Predicate;
