+++ src/main/java/seedu/address/model/order/UniqueOrderList.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.order;
 
 import static java.util.Objects.requireNonNull;
 import static seedu.address.commons.util.CollectionUtil.requireAllNonNull;
@@ -8,8 +8,8 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.order.exceptions.DuplicatePersonException;
+import seedu.address.model.order.exceptions.PersonNotFoundException;
 
 /**
  * A list of orders that enforces uniqueness between its elements and does not allow nulls.
