+++ src/main/java/seedu/address/model/order/exceptions/PersonNotFoundException.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person.exceptions;
+package seedu.address.model.order.exceptions;
 
 /**
  * Signals that the operation is unable to find the specified person.
