+++ src/test/java/seedu/address/logic/commands/AddCommandIntegrationTest.java
@@ -10,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.Order;
 import seedu.address.testutil.OrderBuilder;
 
 /**
