+++ src/test/java/seedu/address/logic/parser/AddCommandParserTest.java
@@ -32,11 +32,11 @@ import static seedu.address.testutil.TypicalPersons.BOB;
 import org.junit.jupiter.api.Test;
 
 import seedu.address.logic.commands.AddCommand;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Order;
-import seedu.address.model.person.Phone;
+import seedu.address.model.order.Address;
+import seedu.address.model.order.Email;
+import seedu.address.model.order.Name;
+import seedu.address.model.order.Order;
+import seedu.address.model.order.Phone;
 import seedu.address.model.tag.Tag;
 import seedu.address.testutil.OrderBuilder;
