+++ src/test/java/seedu/address/logic/parser/OrderParserTest.java
@@ -23,8 +23,8 @@ import seedu.address.logic.commands.FindCommand;
 import seedu.address.logic.commands.HelpCommand;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.NameContainsKeywordsPredicate;
+import seedu.address.model.order.Order;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 import seedu.address.testutil.OrderBuilder;
 import seedu.address.testutil.PersonUtil;
