+++ src/test/java/seedu/address/model/OrderBookTest.java
@@ -18,8 +18,8 @@ import org.junit.jupiter.api.Test;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Order;
-import seedu.address.model.person.exceptions.DuplicatePersonException;
+import seedu.address.model.order.Order;
+import seedu.address.model.order.exceptions.DuplicatePersonException;
 import seedu.address.testutil.OrderBuilder;
 
 public class OrderBookTest {
