+++ src/test/java/seedu/address/model/order/UniqueOrderListTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.order;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -15,8 +15,8 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.order.exceptions.DuplicatePersonException;
+import seedu.address.model.order.exceptions.PersonNotFoundException;
 import seedu.address.testutil.OrderBuilder;
 
 public class UniqueOrderListTest {
