+++ src/test/java/seedu/address/testutil/OrderBookBuilder.java
@@ -1,7 +1,7 @@
 package seedu.address.testutil;
 
 import seedu.address.model.OrderBook;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.Order;
 
 /**
  * A utility class to help with building Orderbook objects.
