+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -16,7 +16,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.model.OrderBook;
-import seedu.address.model.person.Order;
+import seedu.address.model.order.Order;
 
 /**
  * A utility class containing a list of {@code Person} objects to be used in tests.
